+++ src/baldr/edgeinfo.cc
@@ -2,6 +2,8 @@
 #include "baldr/graphconstants.h"
 #include "midgard/elevation_encoding.h"
 
+#include <vector>
+
 using namespace valhalla::baldr;
 
 namespace {
@@ -11,21 +13,17 @@ bool IsNonLiguisticTagValue(char ch) {
   return static_cast<TaggedValue>(ch) != TaggedValue::kLinguistic;
 }
 
-json::MapPtr bike_network_json(uint8_t mask) {
-  return json::map({
-      {"national", static_cast<bool>(mask & kNcn)},
-      {"regional", static_cast<bool>(mask & kRcn)},
-      {"local", static_cast<bool>(mask & kLcn)},
-      {"mountain", static_cast<bool>(mask & kMcn)},
-  });
+void bike_network_json(uint8_t mask, rapidjson::writer_wrapper_t& writer) {
+  writer("national", static_cast<bool>(mask & kNcn));
+  writer("regional", static_cast<bool>(mask & kRcn));
+  writer("local", static_cast<bool>(mask & kLcn));
+  writer("mountain", static_cast<bool>(mask & kMcn));
 }
 
-json::ArrayPtr names_json(const std::vector<std::string>& names) {
-  auto a = json::array({});
+void names_json(const std::vector<std::string>& names, rapidjson::writer_wrapper_t& writer) {
   for (const auto& n : names) {
-    a->push_back(n);
+    writer(n);
   }
-  return a;
 }
 
 /**
@@ -515,81 +513,78 @@ std::vector<std::string> EdgeInfo::level_ref() const {
   return values;
 }
 
-json::MapPtr EdgeInfo::json() const {
-  json::MapPtr edge_info = json::map({
-      {"way_id", static_cast<uint64_t>(wayid())},
-      {"bike_network", bike_network_json(bike_network())},
-      {"names", names_json(GetNames())},
-      {"shape", midgard::encode(shape())},
-  });
+void EdgeInfo::json(rapidjson::writer_wrapper_t& writer) const {
+  writer("way_id", static_cast<uint64_t>(wayid()));
+
+  writer.start_object("bike_network");
+  bike_network_json(bike_network(), writer);
+  writer.end_object();
+
+  writer.start_array("names");
+  names_json(GetNames(), writer);
+  writer.end_array();
+
+  writer("shape", midgard::encode(shape()));
+
   // add the mean_elevation depending on its validity
   const auto elev = mean_elevation();
   if (elev == kNoElevationData) {
-    edge_info->emplace("mean_elevation", nullptr);
+    writer("mean_elevation", nullptr);
   } else {
-    edge_info->emplace("mean_elevation", static_cast<int64_t>(elev));
+    writer("mean_elevation", static_cast<int64_t>(elev));
   }
 
   if (speed_limit() == kUnlimitedSpeedLimit) {
-    edge_info->emplace("speed_limit", std::string("unlimited"));
+    writer("speed_limit", "unlimited");
   } else {
-    edge_info->emplace("speed_limit", static_cast<uint64_t>(speed_limit()));
+    writer("speed_limit", static_cast<uint64_t>(speed_limit()));
   }
 
-  json::MapPtr conditional_speed_limits;
+  std::vector<std::vector<std::pair<float, float>>> levels;
   for (const auto& [tag, value] : GetTags()) {
-    switch (tag) {
-      case TaggedValue::kLayer:
-        break;
-      case TaggedValue::kLinguistic:
-        break;
-      case TaggedValue::kBssInfo:
-        break;
-      case TaggedValue::kLevel:
-        break;
-      case TaggedValue::kLevelRef:
-        break;
-      case TaggedValue::kLandmark:
-        break;
-      case TaggedValue::kLevels: {
-        json::ArrayPtr levels = json::array({});
-        std::vector<std::pair<float, float>> decoded;
-        uint32_t precision;
-        std::tie(decoded, precision) = decode_levels(value);
-        for (auto& range : decoded) {
-          if (range.first == range.second) {
-            // single number
-            levels->emplace_back(json::fixed_t{range.first, precision});
-          } else {
-            // range
-            json::ArrayPtr level = json::array({});
-            level->emplace_back(json::fixed_t{range.first, precision});
-            level->emplace_back(json::fixed_t{range.second, precision});
-            levels->emplace_back(level);
-          }
-        }
-        edge_info->emplace("levels", levels);
-        break;
+    if (tag == TaggedValue::kLevels) {
+      std::vector<std::pair<float, float>> decoded;
+      uint32_t precision;
+      std::tie(decoded, precision) = decode_levels(value);
+      if (decoded.size()) {
+        levels.push_back(decoded);
       }
-      case TaggedValue::kConditionalSpeedLimits: {
-        if (!conditional_speed_limits) {
-          conditional_speed_limits = json::map({});
-        }
-        const ConditionalSpeedLimit* l = reinterpret_cast<const ConditionalSpeedLimit*>(value.data());
-        conditional_speed_limits->emplace(l->td_.to_string(), static_cast<uint64_t>(l->speed_));
-        break;
-      }
-      case TaggedValue::kTunnel:
-        break;
-      case TaggedValue::kBridge:
-        break;
     }
   }
-  if (conditional_speed_limits) {
-    edge_info->emplace("conditional_speed_limits", std::move(conditional_speed_limits));
+
+  if (levels.size()) {
+    writer.start_array("levels");
+    // precision variable causes issues in the json writer
+    writer.set_precision(3);
+    for (auto& level : levels) {
+      for (auto& range : level) {
+        if (range.first == range.second) {
+          writer(range.first);
+        } else {
+          writer.start_array();
+          writer(range.first);
+          writer(range.second);
+          writer.end_array();
+        }
+      }
+    }
+    writer.end_array();
   }
 
-  return edge_info;
+  std::vector<std::pair<std::string, uint64_t>> keyToValSpeedLimits;
+  for (const auto& [tag, value] : GetTags()) {
+    if (tag == TaggedValue::kConditionalSpeedLimits) {
+      const ConditionalSpeedLimit* l = reinterpret_cast<const ConditionalSpeedLimit*>(value.data());
+      keyToValSpeedLimits.push_back({l->td_.to_string(), l->speed_});
+    }
+  }
+  if (keyToValSpeedLimits.size()) {
+    writer.start_object("conditional_speed_limits");
+    for (auto& p : keyToValSpeedLimits) {
+      writer(p.first, static_cast<uint64_t>(p.second));
+    }
+    writer.end_object();
+  }
 }
 
 } // namespace baldr