+++ valhalla/baldr/traffictile.h
@@ -9,7 +9,8 @@
 #ifndef C_ONLY_INTERFACE
 #include <valhalla/baldr/graphconstants.h>
 #include <valhalla/baldr/graphmemory.h>
-#include <valhalla/baldr/json.h>
+#include <valhalla/baldr/rapidjson_utils.h>
+#include <valhalla/valhalla.h>
 
 #include <algorithm>
 #include <cstdint>
@@ -20,7 +21,6 @@
 #else
 #include <stdint.h>
 #endif
-#include <valhalla/valhalla.h>
 
 #ifndef C_ONLY_INTERFACE
 namespace valhalla {
@@ -137,46 +137,55 @@ struct TrafficSpeed {
         congestion3{c3}, has_incidents{incidents}, spare{0} {
   }
 
-  json::MapPtr json() const volatile {
-    auto live_speed = json::map({});
+  void json(rapidjson::writer_wrapper_t& writer) const volatile {
     if (speed_valid()) {
-      live_speed->emplace("overall_speed", static_cast<uint64_t>(get_overall_speed()));
+      writer("overall_speed", static_cast<uint64_t>(get_overall_speed()));
       auto speed = static_cast<uint64_t>(get_speed(0));
       if (speed == UNKNOWN_TRAFFIC_SPEED_KPH)
-        live_speed->emplace("speed_0", nullptr);
+        writer("speed_0", nullptr);
       else
-        live_speed->emplace("speed_0", speed);
+        writer("speed_0", speed);
       auto congestion = (congestion1 - 1.0) / 62.0;
       if (congestion < 0)
-        live_speed->emplace("congestion_0", nullptr);
-      else
-        live_speed->emplace("congestion_0", json::fixed_t{congestion, 2});
-      live_speed->emplace("breakpoint_0", json::fixed_t{breakpoint1 / 255.0, 2});
+        writer("congestion_0", nullptr);
+      else {
+        writer.set_precision(2);
+        writer("congestion_0", congestion);
+      }
+      writer.set_precision(2);
+      writer("breakpoint_0", breakpoint1 / 255.0);
+      writer.set_precision(3);
 
       speed = static_cast<uint64_t>(get_speed(1));
       if (speed == UNKNOWN_TRAFFIC_SPEED_KPH)
-        live_speed->emplace("speed_1", nullptr);
+        writer("speed_1", nullptr);
       else
-        live_speed->emplace("speed_1", speed);
+        writer("speed_1", speed);
       congestion = (congestion2 - 1.0) / 62.0;
       if (congestion < 0)
-        live_speed->emplace("congestion_1", nullptr);
-      else
-        live_speed->emplace("congestion_1", json::fixed_t{congestion, 2});
-      live_speed->emplace("breakpoint_1", json::fixed_t{breakpoint2 / 255.0, 2});
+        writer("congestion_1", nullptr);
+      else {
+        writer.set_precision(2);
+        writer("congestion_1", congestion);
+      }
+      writer.set_precision(2);
+      writer("breakpoint_1", breakpoint2 / 255.0);
+      writer.set_precision(3);
 
       speed = static_cast<uint64_t>(get_speed(2));
       if (speed == UNKNOWN_TRAFFIC_SPEED_KPH)
-        live_speed->emplace("speed_2", nullptr);
+        writer("speed_2", nullptr);
       else
-        live_speed->emplace("speed_2", speed);
+        writer("speed_2", speed);
       congestion = (congestion3 - 1.0) / 62.0;
       if (congestion < 0)
-        live_speed->emplace("congestion_2", nullptr);
-      else
-        live_speed->emplace("congestion_2", json::fixed_t{congestion, 2});
+        writer("congestion_2", nullptr);
+      else {
+        writer.set_precision(2);
+        writer("congestion_2", congestion);
+        writer.set_precision(3);
+      }
     }
-    return live_speed;
   }
 #endif
 };